+++ test/dummy/boot_check.rb
@@ -9,5 +9,6 @@ Rails.application.eager_load!
 abort "engine is not isolated" unless SolidObjects::Engine.isolated?
 abort "record is not loaded" unless SolidObjects::Record < ActiveRecord::Base
 abort "actor helper is not installed" unless ActionView::Base < SolidObjects::ActorHelper
+abort "actor channel is not loaded" unless "SolidObjects::ActorChannel".safe_constantize
 
 puts "solid_objects_dummy_booted"